+++ quil-py/src/instruction/gate.rs
@@ -21,8 +21,8 @@ use rigetti_pyo3::{
 use strum;
 
 use crate::{
-    expression::PyExpression, impl_copy_for_instruction, impl_eq, impl_to_quil,
-    instruction::PyQubit,
+    expression::PyExpression, impl_copy_for_instruction, impl_eq, impl_pickle_for_instruction,
+    impl_to_quil, instruction::PyQubit,
 };
 
 wrap_error!(RustGateError(quil_rs::instruction::GateError));
@@ -32,7 +32,7 @@ py_wrap_error!(quil, RustParseEnumError, EnumParseError, PyValueError);
 
 py_wrap_data_struct! {
     #[derive(Debug, PartialEq, Eq)]
-    #[pyo3(subclass)]
+    #[pyo3(subclass, module = "quil.instructions")]
     PyGate(Gate) as "Gate" {
         name: String => Py<PyString>,
         parameters: Vec<Expression> => Vec<PyExpression>,
@@ -45,6 +45,7 @@ impl_copy_for_instruction!(PyGate);
 impl_to_quil!(PyGate);
 impl_hash!(PyGate);
 impl_eq!(PyGate);
+impl_pickle_for_instruction!(PyGate);
 
 #[pymethods]
 impl PyGate {
@@ -259,7 +260,7 @@ impl_eq!(PyGateSpecification);
 
 py_wrap_data_struct! {
     #[derive(Debug, PartialEq, Eq)]
-    #[pyo3(subclass)]
+    #[pyo3(subclass, module = "quil.instructions")]
     PyGateDefinition(GateDefinition) as "GateDefinition" {
         name: String => Py<PyString>,
         parameters: Vec<String> => Vec<Py<PyString>>,
@@ -271,6 +272,7 @@ impl_to_quil!(PyGateDefinition);
 impl_copy_for_instruction!(PyGateDefinition);
 impl_hash!(PyGateDefinition);
 impl_eq!(PyGateDefinition);
+impl_pickle_for_instruction!(PyGateDefinition);
 
 #[pymethods]
 impl PyGateDefinition {