+++ quil-py/src/instruction/reset.rs
@@ -6,11 +6,14 @@ use rigetti_pyo3::{
     PyTryFrom,
 };
 
-use crate::{impl_copy_for_instruction, impl_eq, impl_to_quil, instruction::PyQubit};
+use crate::{
+    impl_copy_for_instruction, impl_eq, impl_pickle_for_instruction, impl_to_quil,
+    instruction::PyQubit,
+};
 
 py_wrap_data_struct! {
     #[derive(Debug, PartialEq, Eq)]
-    #[pyo3(subclass)]
+    #[pyo3(subclass, module = "quil.instructions")]
     PyReset(Reset) as "Reset" {
         qubit: Option<Qubit> => Option<PyQubit>
     }
@@ -20,6 +23,7 @@ impl_copy_for_instruction!(PyReset);
 impl_to_quil!(PyReset);
 impl_hash!(PyReset);
 impl_eq!(PyReset);
+impl_pickle_for_instruction!(PyReset);
 
 #[pymethods]
 impl PyReset {