+++ quil-py/src/instruction/timing.rs
@@ -8,11 +8,12 @@ use rigetti_pyo3::{
 };
 
 use super::PyQubit;
+use crate::impl_pickle_for_instruction;
 use crate::{expression::PyExpression, impl_copy_for_instruction, impl_eq, impl_to_quil};
 
 py_wrap_data_struct! {
     #[derive(Debug, PartialEq, Eq)]
-    #[pyo3(subclass)]
+    #[pyo3(subclass, module = "quil.instructions")]
     PyDelay(Delay) as "Delay" {
         duration: Expression => PyExpression,
         frame_names: Vec<String> => Vec<Py<PyString>>,
@@ -24,6 +25,7 @@ impl_to_quil!(PyDelay);
 impl_copy_for_instruction!(PyDelay);
 impl_hash!(PyDelay);
 impl_eq!(PyDelay);
+impl_pickle_for_instruction!(PyDelay);
 
 #[pymethods]
 impl PyDelay {
@@ -44,7 +46,7 @@ impl PyDelay {
 
 py_wrap_data_struct! {
     #[derive(Debug, PartialEq, Eq)]
-    #[pyo3(subclass)]
+    #[pyo3(subclass, module = "quil.instructions")]
     PyFence(Fence) as "Fence" {
         qubits: Vec<Qubit> => Vec<PyQubit>
     }
@@ -54,6 +56,7 @@ impl_to_quil!(PyFence);
 impl_copy_for_instruction!(PyFence);
 impl_hash!(PyFence);
 impl_eq!(PyFence);
+impl_pickle_for_instruction!(PyFence);
 
 #[pymethods]
 impl PyFence {